Add root wrapper
